requirements.txt
google-adk==1.14.1
python-dotenv==1.1.0